requirements.txt
streamlit==1.25.0
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scikit-learn-extra==0.3.0
umap-learn==0.5.5
langchain==0.0.267
langchain-community
pypdf2==3.0.1
groq==0.23.1
plotly==5.15.0
sqlite3
altair==4.2.2
